extract ordered locations method to module
