+++ spec/models/subject_spec.rb
@@ -82,36 +82,12 @@ describe Subject, :type => :model do
     end
   end
 
-  describe "#ordered_locations" do
-    let(:subject) do
-      create(:subject, :with_mediums, :with_subject_sets, num_sets: 1)
-    end
-
-    it "should sort the loaded locations by index" do
-      expected = subject.locations.sort_by { |loc| loc.metadata["index"] }
-      expect(expected.map(&:id)).to eq(subject.ordered_locations.map(&:id))
-    end
-
-    it "should sort the non-loaded locations by db index" do
-      lone_subject = Subject.find(subject.id)
-      expect_any_instance_of(Medium::ActiveRecord_Associations_CollectionProxy)
-        .to receive(:order)
-        .with("\"media\".\"metadata\"->>'index' ASC")
-        .and_call_original
-      expected = subject.locations.sort_by { |loc| loc.metadata["index"] }
-      expect(expected.map(&:id)).to eq(lone_subject.ordered_locations.map(&:id))
-    end
-
-    context "subject without location metadata" do
-      before do
-        subject.locations.update_all(metadata: nil)
-        subject.locations(true)
+  describe "ordered_locations" do
+    it_behaves_like "it has ordered locations" do
+      let(:resource) do
+        create(:subject, :with_mediums, :with_subject_sets, num_sets: 1)
       end
-
-      it "should mimic the database order by using the relation ordering" do
-        expected = subject.locations.order("\"media\".\"metadata\"->>'index' ASC")
-        expect(expected.map(&:id)).to eq(subject.ordered_locations.map(&:id))
-      end
+      let(:klass) { Subject }
     end
   end
